+++ CHANGELOG.md
@@ -134,6 +134,7 @@
 
 ## 1.5.1
 - `Mgmg::Recipe`を実装し，レシピ`String`または`Enumerable`と，注目パラメータ`Symbol`，オプション`Mgmg::Option`をセットで取り扱えるようにした．
-	- `String#to_recipe(para=:power, **kw)`または`Enumerable#to_recipe(para=:power, **kw)`で生成できる．
+	- `String#to_recipe(para=:power, allow_over20: false, **kw)`または`Enumerable#to_recipe(para=:power, **kw)`で生成できる．
 	- `Mgmg::Recipe#build`，`Mgmg::Recipe#search`など`String`等に対する操作と同様にでき，注目パラメータとオプションは，`to_recipe`でセットしたものがデフォルトで使われるようになる．各メソッド呼び出し時にキーワード引数を与えることで，一時的に上書きすることもできる．
+- `String#to_recipe`にのみ，☆20制限のチェック機構を導入した．
 - 計算を繰り返した際，複数装備の装備数が増加していってしまうバグを修正した．